Second Idea Blocked Out (Unfinished)
Started work on my second idea as I'm wanting to have a fall back option to go to.
Based off of 'ZIP-IT', an adventure center in Boyle, Roscommon. Has an escape room and a ziplining obstacle course in the trees. I combined the two of the activities into one obstacle course here. Player starts in the escape room before moving onto the zipline course in the trees